Move basic version command to fluvio-cli-common
