+++ crates/cdk/Cargo.toml
@@ -28,12 +28,11 @@ comfy-table = { workspace = true  }
 sysinfo = { workspace = true, default-features = false }
 cargo-generate = { workspace = true }
 include_dir = { workspace = true }
-sha2 = { workspace = true }
 tempfile = { workspace = true }
 current_platform = { workspace = true }
 
 fluvio = { workspace = true }
-fluvio-cli-common = { workspace = true }
+fluvio-cli-common = { workspace = true, features = ["version-cmd"] }
 fluvio-connector-deployer = { path = "../fluvio-connector-deployer"}
 fluvio-connector-package = { workspace = true,  features = ["toml"]}
 fluvio-future = { workspace = true, features = ["subscriber"]}
