+++ crates/cdk/src/cmd.rs
@@ -1,6 +1,7 @@
+use anyhow::Result;
 use cargo_builder::package::PackageOption;
 use clap::Parser;
-use anyhow::Result;
+use fluvio_cli_common::version_cmd::BasicVersionCmd;
 
 use crate::build::BuildCmd;
 use crate::generate::GenerateCmd;
@@ -8,7 +9,6 @@ use crate::deploy::DeployCmd;
 use crate::test::TestCmd;
 use crate::publish::PublishCmd;
 use crate::set_public::SetPublicCmd;
-use crate::version::VersionCmd;
 
 /// Connector Development Kit
 #[derive(Debug, Parser)]
@@ -18,7 +18,7 @@ pub enum CdkCommand {
     Generate(GenerateCmd),
     Deploy(DeployCmd),
     Publish(PublishCmd),
-    Version(VersionCmd),
+    Version(BasicVersionCmd),
 
     #[command(name = "set-public")]
     SetPublic(SetPublicCmd),
@@ -33,8 +33,7 @@ impl CdkCommand {
             CdkCommand::Publish(opt) => opt.process(),
             CdkCommand::SetPublic(opt) => opt.process(),
             CdkCommand::Generate(opt) => opt.process(),
-            CdkCommand::Version(opt) => opt.process(),
-            
+            CdkCommand::Version(opt) => opt.process("CDK"),
         }
     }
 }
