+++ crates/cdk/src/main.rs
@@ -7,7 +7,6 @@ mod publish;
 mod set_public;
 
 pub(crate) mod utils;
-mod version;
 
 fn main() -> anyhow::Result<()> {
     use clap::Parser;
