+++ crates/smartmodule-development-kit/Cargo.toml
@@ -25,8 +25,6 @@ dirs = { workspace = true }
 toml = { workspace = true }
 cargo-generate = { workspace = true }
 include_dir = { workspace = true }
-sha2 = { workspace = true }
-sysinfo = { workspace = true }
 tempfile = { workspace = true }
 enum-display = "0.1.3"
 lib-cargo-crate = "0.2.1"
@@ -41,5 +39,5 @@ fluvio-extension-common = { path = "../fluvio-extension-common", features = ["ta
 fluvio-smartmodule = { path = "../fluvio-smartmodule",  default-features = false }
 fluvio-controlplane-metadata = { path = "../fluvio-controlplane-metadata", features = ["smartmodule"] }
 fluvio-sc-schema = { path = "../fluvio-sc-schema" }
-fluvio-cli-common = { path = "../fluvio-cli-common", features = ["file-records"]}
+fluvio-cli-common = { path = "../fluvio-cli-common", features = ["file-records", "version-cmd"]}
 cargo-builder = { path = "../cargo-builder"}
