+++ crates/smartmodule-development-kit/src/cmd.rs
@@ -1,6 +1,7 @@
+use anyhow::Result;
 use cargo_builder::package::PackageOption;
 use clap::Parser;
-use anyhow::Result;
+use fluvio_cli_common::version_cmd::BasicVersionCmd;
 
 use crate::build::{BuildCmd, BUILD_TARGET};
 use crate::generate::GenerateCmd;
@@ -9,7 +10,6 @@ use crate::load::LoadCmd;
 use crate::publish::PublishCmd;
 use crate::hub::HubCmd;
 use crate::set_public::SetPublicOpt;
-use crate::version::VersionCmd;
 
 /// SmartModule Development Kit utility
 #[derive(Debug, Parser)]
@@ -29,7 +29,7 @@ pub enum SmdkCommand {
     #[command(name = "set-public")]
     SetPublic(SetPublicOpt),
     /// Print smdk version information
-    Version(VersionCmd),
+    Version(BasicVersionCmd),
 }
 
 impl SmdkCommand {
@@ -42,7 +42,7 @@ impl SmdkCommand {
             SmdkCommand::Publish(opt) => opt.process(),
             SmdkCommand::Hub(opt) => opt.process(),
             SmdkCommand::SetPublic(opt) => opt.process(),
-            SmdkCommand::Version(opt) => opt.process(),
+            SmdkCommand::Version(opt) => opt.process("SMDK"),
         }
     }
 }
