+++ crates/smartmodule-development-kit/src/main.rs
@@ -6,7 +6,6 @@ mod load;
 mod publish;
 mod hub;
 mod set_public;
-mod version;
 
 use std::path::PathBuf;
